Rename `payment.model` to `payment`
